Allow Handler to create resources
A Handler is now a description, whilst the actual handling is carried
out by a RouteHandler. This allows a Handler to construct Resources when
it builds a RouteHandler. This will be used by the asset pipeline to
construct file watchers.

This requires the Runtime be factored into two parts, one available
before the http4s server is constructed, and one available after it is
constructed.