Asset serving (#31)
Create an `AssetRoute`, which watches a directory for changes, and serves assets with a name constructed to include a hash. This prevents browser caching of files that are frequently changing, such as CSS.

The current design uses a key into the `KropRuntime`. This is likely to change as patterns for creating resources stabilise.

A Handler is now a description, whilst the actual handling is carried
out by a RouteHandler. This allows a Handler to construct Resources when
it builds a RouteHandler. This will be used by the asset pipeline to
construct file watchers.

This requires the Runtime be factored into two parts, one available
before the http4s server is constructed, and one available after it is
constructed.

We require at least JDK 17. This is caused by our dependency on HexFormat, but Scala is moving
towards this as a minimum JDK dependency so it shouldn't cause us a problem.